Fix potential HTML ID collision of content-action
The content action itself does not need an ID. The same ID is also used
for a link within the "book chapter navigation panel".

While it is not very likely that they may actually collide (`NS_BOOK` page
is usually not part of a book iself).

This will also fix an issue with a missing "edit" content-action on
`REL1_39`, which causes pages in `NS_BOOK` to break.

ERM29488

[4.2.x]

Change-Id: If1be411e6a88b19209cedbcbf4758b300257e509